Drop dependency on postgres-tmp
It is not possibly to have finalizers that are always run in hspec as
far as I can tell. The reason for this is that the threads with the
finalizers die immediately when the main thread dies and afaik there is
no way to get hspec to wait for the threads to terminate.